improve GrpcServletApplication interface and move it to new class
